feat: :sparkles: Adminpanel changes, Profile-page, User-API routes with Auth, fixing useEffect hooks, Header-styling added
Styling on header-fixed. Admin-panel is done being styled(missing basic functionality, lol), User-API routes works, both for your own-data and for data on user-id. Name changes on some useState's to match the rest of the project.

Staff-page and rule-page will be empty, as long as the database is not populated with users/rules