feat: :sparkles: Quals rank-check, notification + admin-panel page to enable/change it, redid the user-search field in the adminpanel
Added a page for a user to check, if they're eligeble to play together, will warn them if not + the amount of ranks they're off by. Alert them if allowed. Update the user-search bar in the admin-panel to fit better/look nicer. Added stats-page with basic stats. Added a page to toggle/edit notification bar at bottom of page + it's visibility.